pyproject.toml
[project]
name = "biquadr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pyinstaller>=6.16.0",
    "pyqt6>=6.9.1",
    "scipy>=1.16.2",
]
